Fixed cast

+++ c/pthread-atomic-qualifier/atomic-qual-cast.c
@@ -14,14 +14,14 @@ extern void abort(void);
 void reach_error() { assert(0); }
 
 void *thr(void *arg) {
-  if ((int)(_Atomic int)41 != 41) reach_error(); // UNREACH
+  if ((_Atomic int)41 != 41) reach_error(); // UNREACH
   return 0;
 }
 
 int main(void) {
   pthread_t id;
   pthread_create(&id, 0, thr, 0);
-  if ((int)(_Atomic(int))42 != 42) reach_error(); // UNREACH
+  if ((_Atomic(int))42 != 42) reach_error(); // UNREACH
   pthread_join(id, 0);
   return 0;
 }
